+++ CHANGELOG.md
@@ -23,6 +23,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -   Fixed crash with processing patterns that have no outputs due to removing mods.
 -   Fixed not being able to shift click items in the Crafting Grid when an External Storage was connected, on NeoForge.
 -   Fixed locking mode on the Autocrafter "Lock until connected machine is empty" inserting too many resources when using speed upgrades.
+-   Fixed processing and crafting pattern tooltips sometimes not being large enough.
 
 ## [2.0.0-beta.1] - 2025-03-28